+++ ppu/include/sys/poll.h
@@ -12,10 +12,6 @@
 #define POLLWRBAND	0x0100
 #define POLLOUT		POLLWRNORM
 
-#ifdef __cplusplus
-extern "C" {
-#endif
-
 typedef unsigned int nfds_t;
 
 struct pollfd {
@@ -23,6 +19,10 @@ struct pollfd {
 	short events;
 	short revents;
 };
+
+#ifdef __cplusplus
+extern "C" {
+#endif
 
 int poll(struct pollfd fds[],nfds_t nfds,int timeout);
